+++ CHANGELOG.md
@@ -1,3 +1,18 @@
-## 1.0.0
+# 1.1.0
 
-* initial release.
+- Fix bugs
+- All periodic functions now are returning a nullable generic typed Future `Future<T?>`
+- Add `fallback`  in case if that callback future returns null
+  * ```dart
+    Once.runOnEveryNewVersion(
+      callback: () {
+        /* */
+      },
+      fallback: () {
+        /* */
+      },
+    );
+    ```
+# 1.0.0
+
+- initial release.